# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Architecture Control Plane API
  version: 0.1.0
  description: |
    API contract for cycle orchestration, approval, and result retrieval.

    Current scaffold authentication is header-based:
    - `X-User-Id` is required on authenticated endpoints.
    - `X-User-Role` is optional and defaults to `operator`.
    - `X-Tenant-Id` is optional.
    - `X-Request-Id` is optional and is echoed in the response envelope body as `request_id`.
servers:
  - url: /v1
security:
  - UserIdentityHeader: []
...
paths:
  /cycles:
    post:
      tags: [cycles]
      summary: Create a new work cycle
      operationId: createCycle
      security:
        - UserIdentityHeader: []
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XUserRoleHeader'
        - $ref: '#/components/parameters/XTenantIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCycleRequest'
            examples:
              basic:
                value:
                  project_id: project-1
                  user_input: Build hydraulic diagnostic plan
                  metadata:
                    priority: normal
      responses:
        '201':
          description: Cycle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeCreateCycleResponse'
        '200':
          description: Existing idempotent result returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeCreateCycleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /cycles/{cycle_id}:
    get:
      tags: [cycles]
      summary: Get cycle summary
      operationId: getCycle
      security:
        - UserIdentityHeader: []
      parameters:
        - $ref: '#/components/parameters/CycleIdPath'
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XUserRoleHeader'
        - $ref: '#/components/parameters/XTenantIdHeader'
      responses:
        '200':
          description: Cycle summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeCycleSummaryResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    UserIdentityHeader:
      type: apiKey
      in: header
      name: X-User-Id
      description: Required identity header used by the scaffold auth dependency.
  parameters:
    CycleIdPath:
      in: path
      name: cycle_id
      required: true
      schema:
        type: string
      description: Cycle identifier.
...
    XRequestIdHeader:
      in: header
      name: X-Request-Id
      required: false
      schema:
        type: string
      description: Optional request identifier. If omitted, the server generates one and returns it in the response envelope body.
    XUserRoleHeader:
      in: header
      name: X-User-Role
      required: false
      schema:
        type: string
        default: operator
      description: Optional scaffold role header. Used by approval authorization checks.
    XTenantIdHeader:
      in: header
      name: X-Tenant-Id
      required: false
      schema:
        type: string
      description: Optional tenant scope header. When present, tenant ownership checks are enforced.
    IdempotencyKeyHeader:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
      description: Required idempotency key for mutating endpoints.
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            badRequest:
              value:
                request_id: req-123
                correlation_id: corr-123
                status: error
                error:
                  code: bad_request
                  message: malformed request
                  retryable: false
                  next_action: null
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            forbidden:
              value:
                request_id: req-123
                correlation_id: corr-123
                status: error
                error:
                  code: forbidden
                  message: forbidden
                  retryable: false
                  next_action: null
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            notFound:
              value:
                request_id: req-123
                correlation_id: corr-123
                status: error
                error:
                  code: not_found
                  message: cycle not found
                  retryable: false
                  next_action: null
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            conflict:
              value:
                request_id: req-123
                correlation_id: corr-123
                status: error
                error:
                  code: conflict
                  message: idempotency key reused with different request payload
                  retryable: false
                  next_action: null
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            validationError:
              value:
                request_id: req-123
                correlation_id: corr-123
                status: error
                error:
                  code: validation_error
                  message: Input should be a valid string
                  retryable: false
                  next_action: null
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internalError:
              value:
                request_id: req-123
                correlation_id: corr-123
                status: error
                error:
                  code: internal_error
                  message: unexpected server error
                  retryable: true
                  next_action: null
  schemas:
    EnvelopeBase:
      type: object
      required: [request_id, correlation_id, status]
      properties:
        request_id:
          type: string
        correlation_id:
          type: string
        status:
          type: string
          enum: [ok, accepted, error]
    ErrorDetail:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        next_action:
          type: string
          nullable: true
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          required: [error]
          properties:
            error:
              $ref: '#/components/schemas/ErrorDetail'
    ArtifactRef:
      type: object
      required: [artifact_id, artifact_type, uri]
      properties:
        artifact_id:
          type: string
        artifact_type:
          type: string
        uri:
          type: string
        content_type:
          type: string
          nullable: true
    CreateCycleRequest:
      type: object
      required: [project_id, user_input]
      properties:
        project_id:
          type: string
        tenant_id:
          type: string
          nullable: true
        user_input:
          type: string
        input_artifacts:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactRef'
        metadata:
          type: object
          additionalProperties: true
    CreateCycleResponse:
      type: object
      required: [cycle_id, state, user_status, approval_required, created_at]
      properties:
        cycle_id:
          type: string
        state:
          type: string
        user_status:
          type: string
        next_action:
          type: string
          nullable: true
        approval_required:
          type: boolean
        created_at:
          type: string
          format: date-time
    EnvelopeCreateCycleResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/CreateCycleResponse'
    CycleSummaryResponse:
      type: object
      required: [cycle_id, state, user_status, approval_required, retry_allowed, replan_allowed, updated_at]
      properties:
        cycle_id:
          type: string
        state:
          type: string
        user_status:
          type: string
        next_action:
          type: string
          nullable: true
        approval_required:
          type: boolean
        retry_allowed:
          type: boolean
        replan_allowed:
          type: boolean
        updated_at:
          type: string
          format: date-time
    EnvelopeCycleSummaryResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/CycleSummaryResponse'
...